+++ src/Filament/EmailLogin.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Laragear\EmailLogin\Filament;
+
+use Filament\Panel;
+
+class EmailLogin extends Panel
+{
+
+}
